+++ src/frontend/citizen-portal/src/app/components/email-verification-required/email-verification-required.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { ticketTypes } from '@shared/enums/ticket-type.enum';
 import { DisputeRepresentedByLawyer } from 'app/api';
-import { DisputeService, NoticeOfDispute } from 'app/services/dispute.service';
+import { NoticeOfDisputeService, NoticeOfDispute } from 'app/services/notice-of-dispute.service';
 import { ViolationTicketService } from 'app/services/violation-ticket.service';
 
 @Component({
@@ -23,23 +23,23 @@ export class EmailVerificationRequiredComponent implements OnInit {
 
   constructor(
     private route: ActivatedRoute,
-    private disputeService: DisputeService,
+    private noticeOfDisputeService: NoticeOfDisputeService,
     private violationTicketService: ViolationTicketService,
   ) {
     this.route.queryParams.subscribe((params) => {
       this.email = params.email;
       this.token = params.token;
     });
-    this.noticeOfDispute = disputeService.noticeOfDispute;
+    this.noticeOfDispute = noticeOfDisputeService.noticeOfDispute;
   }
 
   ngOnInit() {
     this.ticketType = this.violationTicketService.ticketType;
-    this.countsActions = this.disputeService.getCountsActions(this.noticeOfDispute.dispute_counts);
+    this.countsActions = this.noticeOfDisputeService.getCountsActions(this.noticeOfDispute.dispute_counts);
   }
 
   resendEmail() {
-    this.disputeService.resendVerificationEmail(this.token).subscribe(() => { });
+    this.noticeOfDisputeService.resendVerificationEmail(this.token).subscribe(() => { });
   }
 
   onPrint(): void {
